Разбил main.js на модули
