+++ index.html
@@ -6,7 +6,9 @@
   <link rel="stylesheet" href="css/style.css">
   <link rel="shortcut icon" href="favicon.ico" type="image/x-icon">
   <title>Кексобукинг</title>
-  <script src="js/main.js"></script>
+  <script type="module" src="js/init.js"></script>
+  <script type="module" src="js/util.js"></script>
+  <script type='module' src="js/main.js"></script>
 </head>
 <body>
